+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - PR #232 Update phishing detection to use cudf subword tokenizer
 - PR #237 Cybert streamz memory optimization
 - PR #238 Deprecate CLX subword tokenizer
+- PR #239 Added cybert streamz log size parameter
+- PR #244 Add cybert dataloader
+- PR #247 Allow CuPy 8.x
 
 ## Bug Fixes
 - PR #231 Fix segmentation fault in cybert notebook
@@ -17,6 +20,7 @@
 - PR #240 Fix errors from upstream changes to cuDF DataFrame.drop and RMM
 - PR #241 Fix unit tests after cuDF update
 - PR #242 Pin seqeval to v0.0.12 for cybert example training notebook
+- PR #243 Fix error handling in `ci/gpu/build.sh`
 
 # clx 0.15.0 (26 Aug 2020)